requirements.txt
asgiref==3.7.2
decouple==0.0.7
Django==4.2.7
django-filter==23.4
djangorestframework==3.14.0
importlib-metadata==6.8.0
Markdown==3.5.1
pytz==2023.3.post1
sqlparse==0.4.4
typing-extensions==4.8.0
zipp==3.17.0
asgiref==3.7.2
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
cffi==1.16.0
channels==4.0.0
constantly==23.10.4
cryptography==41.0.7
daphne==4.0.0
Django==4.2.7
django-filter==23.4
djangorestframework==3.14.0
drf-spectacular==0.26.5
hyperlink==21.0.0
idna==3.6
importlib-metadata==6.8.0
incremental==22.10.0
inflection==0.5.1
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
Markdown==3.5.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pyOpenSSL==23.3.0
python-decouple==3.8
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.31.1
rpds-py==0.13.2
service-identity==23.1.0
six==1.16.0
sqlparse==0.4.4
twisted==23.10.0
txaio==23.1.1
typing-extensions==4.8.0
uritemplate==4.1.1
zipp==3.17.0
zope.interface==6.1
